requirements.txt
absl-py==2.0.0
addict==2.4.0
aio-pika==9.5.1
aiofiles==23.2.1
aiormq==6.8.1
altair==5.2.0
annotated-types==0.7.0
anyio==4.6.2.post1
asyncio==3.4.3
attrs==23.1.0
basicsr==1.4.2
blessed==1.20.0
cachetools==5.3.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
distro==1.9.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.105.0
ffmpy==0.3.1
filelock==3.13.1
filetype==1.2.0
filterpy==1.4.5
flatbuffers==24.3.25
fonttools==4.46.0
fsspec==2023.12.2
future==0.18.3
gfpgan==1.3.8
google-auth==2.25.2
google-auth-oauthlib==1.2.0
gradio==3.50.2
gradio_client==0.6.1
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.10
imageio==2.33.1
importlib-metadata==7.0.0
importlib-resources==6.1.1
inquirer==3.1.4
inquirerpy==0.3.4
Jinja2==3.1.2
jiter==0.7.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
Markdown==3.5.1
matplotlib==3.8.2
meshctrl==0.1.15
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.1
oauthlib==3.2.2
onnx==1.15.0
onnxruntime-silicon==1.16.0
openai==1.55.0
opencv-python==4.8.1.78
orjson==3.9.10
pamqp==3.3.0
pandas==2.1.4
pfzy==0.3.4
pika==1.3.2
Pillow==10.1.0
pipenv==2024.0.1
platformdirs==4.1.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==4.25.1
psutil==5.9.6
pyaes==1.6.1
pyasn1==0.6.1
pyasn1-modules==0.3.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.10.1
pydantic_core==2.27.1
pydub==0.25.1
pyparsing==3.1.1
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.1
python-editor==1.0.4
python-multipart==0.0.6
python-telegram-bot==21.7
pytz==2024.2
readchar==4.0.5
realesrgan==0.3.0
referencing==0.32.0
requests==2.32.3
requests-oauthlib==1.3.1
rpds-py==0.15.2
rsa==4.9
scikit-image==0.22.0
scipy==1.11.4
semantic-version==2.10.0
sniffio==1.3.1
starlette==0.27.0
sympy==1.12
tb-nightly==2.16.0a20231217
Telethon==1.38.1
tensorboard-data-server==0.7.2
tf-keras-nightly==2.16.0.dev2023121710
tifffile==2023.12.9
tomli==2.0.1
toolz==0.12.0
torch==2.1.0
torchvision==0.16.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.26.2
wcwidth==0.2.12
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.1
wrapt==1.16.0
yapf==0.40.2
yarl==1.18.0
zipp==3.19.2
